api/requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
billiard==3.6.3.0
celery==4.4.7
certifi==2021.5.30
cffi==1.14.2
charset-normalizer==2.0.4
cryptography==2.9.2
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-nested-routers==0.91
gunicorn==20.0.4
idna==3.2
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
logilab-common==1.8.1
mccabe==0.6.1
mypy-extensions==0.4.3
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
redis==3.5.3
requests==2.26.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.26.6
vine==1.3.0
wrapt==1.12.1
